+++ NEWS.md
@@ -3,7 +3,9 @@
 ## New features
 
 - `fp_par()` now have argument `word_style_id`
-- Add support for `ln` for `external_img` provided by Angus Moore
+- add support for `ln` for `external_img` provided by Angus Moore
+- add word_title and word_description to table properties (`prop_table()`). 
+These values can be used as alternative text for Word tables.
 
 ## Issues